+++ python/gigl/common/data/dataloaders.py
@@ -29,7 +29,8 @@ class SerializedTFRecordInfo:
     tfrecord_uri_prefix: Uri
     # Feature names to load for the current entity
     feature_keys: Sequence[str]
-    # a dict of feature name -> FeatureSpec (eg. FixedLenFeature, VarlenFeature, SparseFeature, RaggedFeature). If entity keys are not present, we insert them during tensor loading
+    # a dict of feature name -> FeatureSpec (eg. FixedLenFeature, VarlenFeature, SparseFeature, RaggedFeature). If entity keys are not present, we insert them during tensor loading.
+    # Note that entity label keys should also be included in the feature_spec if they are present.
     feature_spec: FeatureSpecDict
     # Feature dimension of current entity
     feature_dim: int
@@ -367,10 +368,12 @@ class TFRecordDataLoader:
                 if entity_type == FeatureTypes.NODE
                 else torch.empty(2, 0)
             )
-            if label_keys:
+            if label_keys and feature_keys:
                 empty_feature = torch.empty(
                     0, serialized_tf_record_info.feature_dim + len(label_keys)
                 )
+            elif label_keys and not feature_keys:
+                empty_feature = torch.empty(0, len(label_keys))
             elif feature_keys:
                 empty_feature = torch.empty(0, serialized_tf_record_info.feature_dim)
             else:
